requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
altair
annotated-types
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
babel
bcrypt
beautifulsoup4
bleach
blinker
cachetools
captcha
certifi
cffi
charset-normalizer
click
cloudpickle
comm
cryptography
dask
dataclasses-json
debugpy
decorator
deepmerge
defusedxml
deprecation
distributed
eralchemy2
et-xmlfile
executing
extra-streamlit-components
faiss-cpu
fastjsonschema
fqdn
frozenlist
fsspec
gitdb
gitpython
graphviz
greenlet
h11
h2
hpack
httpcore
httpx
httpx-sse
humanize
hyperframe
idna
importlib-metadata
ipaddress
ipycytoscape
ipykernel
ipython
ipython-pygments-lexers
ipywidgets
isoduration
itables
jedi
jinja2
json5
jsonpatch
jsonpath-ng
jsonpickle
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-ai
jupyter-ai-magics
jupyter-client
jupyter-console
jupyter-core
jupyter-events
jupyter-kernel-gateway
jupyter-lsp
jupyter-server
jupyter-server-terminals
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-vim
jupyterlab-widgets
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
lark
locket
loguru
mako
markupsafe
marshmallow
matplotlib-inline
mistune
msgpack
multidict
mypy-extensions
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook-shim
numpy
openpyxl
orjson
overrides
packaging
pandas
pandocfilters
parso
partd
pexpect
pillow
pip
platformdirs
plotly
ply
postgrest
prometheus-client
prompt-toolkit
propcache
protobuf
psutil
psycopg
psycopg-binary
psycopg2-binary
ptpython
ptyprocess
pure-eval
pyarrow
pycparser
pydantic
pydantic-core
pydantic-settings
pydeck
pygments
pygraphviz
pyjwt
python-dateutil
python-dotenv
python-json-logger
pytz
pyvis
pyyaml
pyzmq
qgrid
realtime
referencing
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rpds-py
send2trash
setuptools
simple-ddl-parser
six
smmap
sniffio
sortedcontainers
soupsieve
spectate
sqlalchemy
stack-data
storage3
streamlit
streamlit-authenticator
strenum
supabase
supabase-auth
supabase-functions
tabulate
tblib
tenacity
terminado
tinycss2
toml
toolz
tornado
traitlets
types-python-dateutil
typing-extensions
typing-inspect
typing-inspection
tzdata
uri-template
urllib3
validators
watchdog
wcwidth
webcolors
webencodings
websocket-client
websockets
widgetsnbextension
xlsxwriter
yarl
zict
zipp
zstandard
